custom_agent_demos/examples/custom_agent/llm_agent/requirements.txt
volcengine-python-sdk[ark]==1.0.112
uvicorn==0.32.1
fastapi==0.115.6
loguru==0.7.2
openai==1.57.0
redis==5.2.0
dashscope==1.20.14
dashvector==1.0.18
pymilvus==2.5.2
oss2==2.19.1
pyyaml==6.0.2